Disable functions when semester is locked

+++ start.py
@@ -75,13 +75,13 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 		section_title = u'Inställningar'
 		baselink += "settings/"
 		breadcrumbs.append({'link':baselink, 'text':section_title})
-		if request.method == "POST":
+		if semester.isOpen() and request.method == "POST":
 			troop.defaultstarttime = request.form['defaultstarttime']
 			troop.defaultduration = int(request.form['defaultduration'])
 			troop.rapportID = int(request.form['rapportID'])
 			troop.put()
 			
-		form = htmlform.HtmlForm('troopsettings')
+		form = htmlform.HtmlForm('troopsettings', disabled=semester.locked)
 		form.AddField('defaultstarttime', troop.defaultstarttime, 'Avdelningens vanliga starttid')
 		form.AddField('defaultduration', troop.defaultduration, u'Avdelningens vanliga mötestid i minuter', 'number')
 		form.AddField('rapportID', troop.rapportID, 'Unik rapport ID för kommunens närvarorapport', 'number')
@@ -93,7 +93,7 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 	if key_url == "delete":
 		if troop == None:
 			return "", 404
-		if len(request.form) > 0 and "confirm" in request.form:
+		if semester.isOpen() and len(request.form) > 0 and "confirm" in request.form:
 			if not user.isGroupAdmin():
 				return "", 403
 			troop.delete()
@@ -121,7 +121,7 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 				breadcrumbs=breadcrumbs,
 				trooppersons=[],
 				scoutgroup=scoutgroup)
-		elif request.method == "POST":
+		elif semester.isOpen() and request.method == "POST":
 			pnr = request.form['personnummer'].replace('-','')
 			person = Person.createlocal(
 				request.form['firstname'], 
@@ -192,7 +192,7 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 			jsonstr+=']'
 			return jsonstr
 		elif action == "addperson":
-			if troop == None or key_url == None:
+			if troop == None or key_url == None or semester.locked:
 				raise ValueError('Missing troop or person')
 			person_key = ndb.Key(urlsafe=key_url)
 			person = person_key.get()
@@ -207,7 +207,7 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 			user.activeSemester = ndb.Key(urlsafe=semester_url)
 			user.put()
 		elif action == "removefromtroop" or action == "setasleader" or action == "removeasleader":
-			if troop == None or key_url == None:
+			if troop == None or key_url == None or semester.locked:
 				raise ValueError('Missing troop or person')
 			person_key = ndb.Key(urlsafe=key_url)
 			tps = TroopPerson.query(TroopPerson.person == person_key, TroopPerson.troop == troop_key).fetch(1)
@@ -226,7 +226,7 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 	if request.method == "POST" and len(request.form) > 0 and "action" in request.form:
 		action=request.form["action"]
 		if action == "saveattendance":
-			if troop == None or scoutgroup == None or key_url == None:
+			if troop == None or scoutgroup == None or key_url == None or semester.locked:
 				raise ValueError('Missing troop or group')
 
 			meeting = ndb.Key(urlsafe=key_url).get()
@@ -239,6 +239,9 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 			meeting.put()
 			return "ok"
 		elif action == "addmeeting" or action == "updatemeeting":
+			if troop == None or scoutgroup == None or key_url == None or semester.locked:
+				raise ValueError('Missing troop or group')
+
 			mname = request.form['name']
 			mdate = request.form['date']
 			mtime = request.form['starttime'].replace('.', ':')
@@ -259,16 +262,24 @@ def show(sgroup_url=None, troop_url=None, key_url=None):
 			meeting.commit()
 			return redirect(breadcrumbs[-1]['link'])
 		elif action == "deletemeeting":
+			if troop == None or scoutgroup == None or key_url == None or semester.locked:
+				raise ValueError('Missing troop or group')
+
 			meeting = ndb.Key(urlsafe=key_url).get()
 			logging.debug("deleting meeting=%s", meeting.getname())
 			meeting.delete()
 			return redirect(breadcrumbs[-1]['link'])
 		elif action == "savepatrol":
+			if troop == None or scoutgroup == None or key_url == None or semester.locked:
+				raise ValueError('Missing troop or group')
+
 			patrolperson = ndb.Key(urlsafe=request.form['person']).get()
 			patrolperson.setpatrol(request.form['patrolName'])
 			patrolperson.put()
 			return "ok"
 		elif action == "newtroop":
+			if semester.locked:
+				raise ValueError('Termin låst')
 			troopname = request.form['troopname']
 			troop_id = hash(troopname)
 			conflict = Troop.get_by_id(Troop.getid(troop_id, scoutgroup.key, user.activeSemester), use_memcache=True)
